requirements.txt
aiogram==3.*
playwright==1.*
beautifulsoup4==4.*
python-dotenv==1.*


